requirements.txt
discord.py[voice]
youtube_dl
cleverbot
requests
pafy
bs4